Crate uran_ecs

Warstwa ECS oparta na hecs, z komponentami pod render 2D.

Re-exports

pub use component::BlendMode;

pub use component::Material;

pub use component::Mesh;

pub use component::Sprite;

pub use component::Transform;

pub use component::UvRect;

pub use component::Vertex;

pub use component::Visibility;

Modules

component

Komponenty renderowania 2D.

Structs

Entity

Lightweight unique ID, or handle, of an entity

QueryBorrow

A borrow of a World sufficient to execute the query Q

Ref

Shared borrow of an entity’s component

RefMut

Unique borrow of an entity’s component

World

An unordered collection of entities, each having any number of distinctly typed components